+++ web.js
@@ -5,9 +5,9 @@
 var config = require('./config');
 var constants = require('./constants.js');
 var utility = require('./util/utility');
-var buildSets = require('./util/buildSets');
+var buildSets = require('./store/buildSets');
 var redis = require('./store/redis');
-var status = require('./store/status');
+var status = require('./store/buildStatus');
 var db = require('./store/db');
 var queries = require('./store/queries');
 var matches = require('./routes/matches');
