+++ worker.js
@@ -7,7 +7,7 @@ var redis = require('./store/redis');
 var queue = require('./store/queue');
 var db = require('./store/db');
 var queries = require('./store/queries');
-var buildSets = require('./util/buildSets');
+var buildSets = require('./store/buildSets');
 var utility = require('./util/utility');
 var getMMStats = require("./util/getMMStats");
 var async = require('async');
